Log in to the aggregator.
Yes, I'm well aware that the credentials are in cleartext in this repo.

A stupid attacker (or someone just casting about; similar idea) is not
likely to check this repo before trying to screw with us.  If it turns
out we need better authorization, we can arrange for that when we need
it.